requirements.txt
absl-py==0.8.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
black==19.3b0
certifi==2019.9.11
cffi==1.12.3
Click==7.0
cryptography==2.7
cycler==0.10.0
dataclasses==0.6
Django==2.2.4
django-rest-framework-mongoengine==3.3.1
django-tastypie==0.14.2
djangorestframework==3.10.3
djongo==1.2.35
Flask==1.1.1
Flask-API==1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
imbalanced-learn==0.5.0
imblearn==0.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mongoengine==0.18.2
nltk==3.4.5
numpy==1.17.1
pandas==0.25.1
pipenv==2018.11.26
protobuf==3.9.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sqlparse==0.2.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toml==0.10.0
virtualenv==16.7.5
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wrapt==1.11.2
